# Do less session updates, remove hash check