--- a/coprocess.go
+++ b/coprocess.go
@@ -324,12 +324,8 @@ func (m *CoProcessMiddleware) ProcessRequest(w http.ResponseWriter, r *http.Requ
 		for k, v := range returnObject.Metadata {
 			returnedSession.MetaData[k] = string(v)
 		}
-		if extractor == nil {
-			ctxSetSession(r, returnedSession, sessionID, true)
-		} else {
-			// The CP middleware did setup a session, we should pass it to the ID extractor (caching):
-			extractor.PostProcess(r, returnedSession, sessionID)
-		}
+
+		ctxSetSession(r, returnedSession, sessionID, true)
 	}
 
 	return nil, 200