--- a/handler_error.go
+++ b/handler_error.go
@@ -31,6 +31,7 @@ type ErrorHandler struct {
 
 // HandleError is the actual error handler and will store the error details in analytics if analytics processing is enabled.
 func (e *ErrorHandler) HandleError(w http.ResponseWriter, r *http.Request, errMsg string, errCode int) {
+	defer e.Base().UpdateRequestSession(r)
 
 	var templateExtension string
 	var contentType string