--- a/mw_api_rate_limit.go
+++ b/mw_api_rate_limit.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/TykTechnologies/tyk/request"
+	"github.com/TykTechnologies/tyk/storage"
 	"github.com/TykTechnologies/tyk/user"
 )
 
@@ -37,6 +38,7 @@ func (k *RateLimitForAPI) EnabledForSpec() bool {
 		Per:         k.Spec.GlobalRateLimit.Per,
 		LastUpdated: strconv.Itoa(int(time.Now().UnixNano())),
 	}
+	k.apiSess.SetKeyHash(storage.HashKey(k.keyName))
 
 	return true
 }