--- a/mw_auth_key.go
+++ b/mw_auth_key.go
@@ -102,8 +102,7 @@ func (k *AuthKey) ProcessRequest(w http.ResponseWriter, r *http.Request, _ inter
 	// Set session state on context, we will need it later
 	switch k.Spec.BaseIdentityProvidedBy {
 	case apidef.AuthToken, apidef.UnsetAuth:
-		ctxSetSession(r, &session)
-		ctxSetAuthToken(r, key)
+		ctxSetSession(r, &session, key, false)
 		k.setContextVars(r, key)
 	}