--- a/mw_basic_auth.go
+++ b/mw_basic_auth.go
@@ -113,8 +113,7 @@ func (k *BasicAuthKeyIsValid) ProcessRequest(w http.ResponseWriter, r *http.Requ
 	// Set session state on context, we will need it later
 	switch k.Spec.BaseIdentityProvidedBy {
 	case apidef.BasicAuthUser, apidef.UnsetAuth:
-		ctxSetSession(r, &session)
-		ctxSetAuthToken(r, keyName)
+		ctxSetSession(r, &session, keyName, false)
 	}
 
 	// Request is valid, carry on