@buger 2018-06-13
@dencoded added change to improve KeyHash caching. Now it should be called only 1 time per session and be cached inside session object.